+++ CHANGES.md
@@ -114,6 +114,19 @@ To be released.
 [Kademlia]: https://en.wikipedia.org/wiki/Kademlia
 
 
+Version 0.5.3
+-------------
+
+Released on September 9, 2019.
+
+### Bug fixes
+
+ -  Fix bug where `IAccountStateDelta.GetState()` hadn't returned proper state
+    when the block action is evaluated.  [[#500]]
+
+[#500]: https://github.com/planetarium/libplanet/pull/500
+
+
 Version 0.5.2
 -------------